emu: add pkg, nonpdrm support
