+++ vita3k/crypto/src/hash.cpp
@@ -25,7 +25,7 @@ extern "C" {
 Sha256Hash sha256(const void *data, size_t size) {
     Sha256Hash hash;
     SHA256_CTX sha_ctx = {};
-    sha256_init(&sha_ctx);
+    sha256_init_one(&sha_ctx);
     sha256_update(&sha_ctx, static_cast<const uint8_t *>(data), size);
     sha256_final(&sha_ctx, hash.data());
