+++ vita3k/gui/src/gui.cpp
@@ -331,10 +331,12 @@ void draw_ui(GuiState &gui, HostState &host) {
 
     ImGui::PushFont(gui.monospaced_font);
 
-    if (gui.file_menu.game_install_dialog)
-        draw_game_install_dialog(gui, host);
     if (gui.file_menu.firmware_install_dialog)
         draw_firmware_install_dialog(gui, host);
+    if (gui.file_menu.pkg_install_dialog)
+        draw_pkg_install_dialog(gui, host);
+    if (gui.file_menu.game_install_dialog)
+        draw_game_install_dialog(gui, host);
     if (gui.debug_menu.threads_dialog)
         draw_threads_dialog(gui, host);
     if (gui.debug_menu.thread_details_dialog)
