+++ vita3k/gui/src/private.h
@@ -41,6 +41,7 @@ const ImVec4 GUI_PROGRESS_BAR = RGBA_TO_FLOAT(118, 150, 55, 255);
 
 void draw_main_menu_bar(GuiState &gui, HostState &host);
 void draw_firmware_install_dialog(GuiState &gui, HostState &host);
+void draw_pkg_install_dialog(GuiState &gui, HostState &host);
 void draw_game_install_dialog(GuiState &gui, HostState &host);
 void draw_threads_dialog(GuiState &gui, HostState &host);
 void draw_thread_details_dialog(GuiState &gui, HostState &host);
