+++ vita3k/host/include/host/pkg.h
@@ -1,0 +1,78 @@
+// Vita3K emulator project
+// Copyright (C) 2020 Vita3K team
+//
+// This program is free software; you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation; either version 2 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License along
+// with this program; if not, write to the Free Software Foundation, Inc.,
+// 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+#include <host/state.h>
+#include <string>
+
+// Credits to mmozeiko https://github.com/mmozeiko/pkg2zip
+
+const uint8_t pkg_vita_2[] = { 0xe3, 0x1a, 0x70, 0xc9, 0xce, 0x1d, 0xd7, 0x2b, 0xf3, 0xc0, 0x62, 0x29, 0x63, 0xf2, 0xec, 0xcb };
+const uint8_t pkg_vita_3[] = { 0x42, 0x3a, 0xca, 0x3a, 0x2b, 0xd5, 0x64, 0x9f, 0x96, 0x86, 0xab, 0xad, 0x6f, 0xd8, 0x80, 0x1f };
+const uint8_t pkg_vita_4[] = { 0xaf, 0x07, 0xfd, 0x59, 0x65, 0x25, 0x27, 0xba, 0xf1, 0x33, 0x89, 0x66, 0x8b, 0x17, 0xd9, 0xea };
+
+enum class PkgType {
+    PKG_TYPE_VITA_APP = 0,
+    PKG_TYPE_VITA_DLC = 1,
+    PKG_TYPE_VITA_PATCH = 2
+};
+
+struct PkgHeader {
+    uint32_t magic;
+    uint16_t revision;
+    uint16_t type;
+    uint32_t info_offset;
+    uint32_t info_count;
+    uint32_t header_size;
+    uint32_t file_count;
+    uint64_t total_size;
+    uint64_t data_offset;
+    uint64_t data_size;
+    char content_id[0x30];
+    uint8_t digest[0x10];
+    uint8_t pkg_data_iv[0x10];
+    uint8_t pkg_signatures[0x40];
+};
+
+struct PkgExtHeader {
+    uint32_t magic;
+    uint32_t unknown_01;
+    uint32_t header_size;
+    uint32_t data_size;
+    uint32_t data_offset;
+    uint32_t data_type;
+    uint64_t pkg_data_size;
+
+    uint32_t padding_01;
+    uint32_t data_type2;
+    uint32_t unknown_02;
+    uint32_t padding_02;
+    uint64_t padding_03;
+    uint64_t padding_04;
+};
+
+struct PkgEntry {
+    uint32_t name_offset;
+    uint32_t name_size;
+    uint64_t data_offset;
+    uint64_t data_size;
+    uint32_t type;
+    uint32_t padding;
+};
+
+bool install_pkg(const std::string &pkg, const std::string &pref_path, std::string &p_zRIF);
+
+bool decrypt_install_nonpdrm(std::string &drmlicpath, const std::string &title_path);
