+++ vita3k/host/src/load_self.cpp
@@ -15,9 +15,9 @@
 // with this program; if not, write to the Free Software Foundation, Inc.,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-#include <host/load_self.h>
 #include <app/functions.h>
 #include <config/state.h>
+#include <host/load_self.h>
 #include <kernel/relocation.h>
 #include <kernel/state.h>
 #include <kernel/types.h>
@@ -342,12 +342,12 @@ SceUID load_self(Ptr<const void> &entry_point, KernelState &kernel, MemState &me
         LOG_CRITICAL("SELF {} header type {} is not supported.", self_path, self_header.header_type);
         return -1;
     }
-    
+
     if (self_path == "app0:sce_module/steroid.suprx") {
         app::error_dialog("You're trying to load a vitamin dump. It is not supported.", nullptr);
         return -1;
     }
-    
+
     const uint8_t *const elf_bytes = self_bytes + self_header.elf_offset;
     const Elf32_Ehdr &elf = *reinterpret_cast<const Elf32_Ehdr *>(elf_bytes);
     const uint32_t module_info_offset = elf.e_entry & 0x3fffffff;
